NetworkPkg: Add support for TLS cipher configuration
This adds support of TLS cipher configuration to the TlsAuthConfigDxe
driver. The driver now allows to configure the list of supported
ciphers and manages the UEFI variable HttpTlsCipherList.

In the UI a new menu is created ("TLS Ciphers Configuration") and
checkboxes for each available cipher.

The list of available ciphers can be selected with the PCD
PcdAvailableTLSCiphers. Note that available does not mean supported.
The TLS library and the underlying crypto library must support the
selected cipher.

The default ciphers can be configured with the PCD PcdDefaultTLSCiphers.
This PCD is also used to initialize the HttpTlsCipherList variable if
it does not exist and the PCD PcdInitializeTLSCipherVariable is set to
TRUE. If some other code is supposed to create this variable, the PCD
can be set to FALSE.

Signed-off-by: Sebastian Witt <sebastian.witt@siemens.com>